Updated style and fixed authError handler

+++ projects/03_coffee_shop_full_stack/starter_code/backend/src/api.py
@@ -18,7 +18,7 @@ CORS(app)
 !! NOTE THIS MUST BE UNCOMMENTED ON FIRST RUN
 !! Running this funciton will add one
 '''
-# db_drop_and_create_all()
+db_drop_and_create_all()
 
 # ROUTES
 '''
@@ -26,9 +26,12 @@ CORS(app)
     GET /drinks
         it should be a public endpoint
         it should contain only the drink.short() data representation
-    returns status code 200 and json {"success": True, "drinks": drinks} where drinks is the list of drinks
+    returns status code 200 and json {"success": True, "drinks": drinks}
+    where drinks is the list of drinks
         or appropriate status code indicating reason for failure
 '''
+
+
 @app.route('/drinks')
 def get_drinks():
     try:
@@ -49,25 +52,33 @@ def get_drinks():
     GET /drinks-detail
         it should require the 'get:drinks-detail' permission
         it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drinks} where drinks is the list of drinks
+    returns status code 200 and json {"success": True, "drinks": drinks}
+    where drinks is the list of drinks
         or appropriate status code indicating reason for failure
 '''
+
+
 @app.route('/drinks-detail')
 @requires_auth('get:drinks-detail')
 def drink_details(payload):
-    try:
-        long_drinks = []
-        drinks = Drink.query.order_by(Drink.id).all()
-        for drink in drinks:
-            long_drinks.append(drink.long())
-        return jsonify({
-            "success": True,
-            "drinks": long_drinks
-        })
 
-    except:
-        print(sys.exc_info())
-        abort(400)
+    if type(payload) is AuthError:
+        return auth_error(payload)
+
+    else:
+        try:
+            long_drinks = []
+            drinks = Drink.query.order_by(Drink.id).all()
+            for drink in drinks:
+                long_drinks.append(drink.long())
+            return jsonify({
+                "success": True,
+                "drinks": long_drinks
+            })
+
+        except:
+            print(sys.exc_info())
+            abort(401)
 
 '''
 @TODO implement endpoint
@@ -75,38 +86,45 @@ def drink_details(payload):
         it should create a new row in the drinks table
         it should require the 'post:drinks' permission
         it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the newly created drink
+    returns status code 200 and json {"success": True, "drinks": drink}
+    where drink an array containing only the newly created drink
         or appropriate status code indicating reason for failure
 '''
+
+
 @app.route('/drinks', methods=['POST'])
 @requires_auth('post:drinks')
 def post_drink(payload):
-    try:
-        # Create a new drink
-        body = request.get_json()
-        drinkTitle = body['title']
-        drinkRecipe = body['recipe']
+    if type(payload) is AuthError:
+        return auth_error(payload)
 
-        # Create and insert the new drink
-        newDrink = Drink(
-            title=drinkTitle,
-            recipe=json.dumps(drinkRecipe)
-        )
-        newDrink.insert()
+    else:
+        try:
+            # Create a new drink
+            body = request.get_json()
+            drinkTitle = body['title']
+            drinkRecipe = body['recipe']
 
-        drinkSelection = Drink.query.order_by(Drink.id).all()
-        newDrinkSelection = []
+            # Create and insert the new drink
+            newDrink = Drink(
+                title=drinkTitle,
+                recipe=json.dumps(drinkRecipe)
+            )
+            newDrink.insert()
 
-        for drink in drinkSelection:
-            newDrinkSelection.append(drink.long())
-        return jsonify({
-            "success": True,
-            "drinks": newDrinkSelection
-        })
+            drinkSelection = Drink.query.order_by(Drink.id).all()
+            newDrinkSelection = []
 
-    except:
-        print(sys.exc_info())
-        abort(400)
+            for drink in drinkSelection:
+                newDrinkSelection.append(drink.long())
+            return jsonify({
+                "success": True,
+                "drinks": newDrinkSelection
+            })
+
+        except:
+            print(sys.exc_info())
+            abort(400)
 
 '''
 @TODO implement endpoint
@@ -116,40 +134,50 @@ def post_drink(payload):
         it should update the corresponding row for <id>
         it should require the 'patch:drinks' permission
         it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the updated drink
+    returns status code 200 and json {"success": True, "drinks": drink}
+    where drink an array containing only the updated drink
         or appropriate status code indicating reason for failure
 '''
+
+
 @app.route('/drinks/<int:drink_id>', methods=['PATCH'])
 @requires_auth('patch:drinks')
 def edit_drink(payload, drink_id):
-    try:
-        body = request.get_json()
-        drink = Drink.query.get(drink_id)
+    if type(payload) is AuthError:
+        return auth_error(payload)
+    else:
+        try:
+            body = request.get_json()
+            drink = Drink.query.get(drink_id)
 
-        if not drink:
-            abort(404)
+            if not drink:
+                abort(404)
 
-        # Get the json body to update the drink
-        drinkTitle = body['title']
-        drinkRecipe = body['recipe']
-        drink.recipe = json.dumps(drinkRecipe)
-        drink.title = drinkTitle
-        drink.update()
+            # Get the json body to update the drink
+            if 'title' in body:
+                drinkTitle = body['title']
+                drink.title = drinkTitle
 
-        # Get the updated drink selection
-        drinkSelection = Drink.query.order_by(Drink.id).all()
-        newDrinkSelection = []
+            if 'recipe' in body:
+                drinkRecipe = body['recipe']
+                drink.recipe = json.dumps(drinkRecipe)
 
-        for drink in drinkSelection:
-            newDrinkSelection.append(drink.long())
+            drink.update()
 
-        return jsonify({
-            "success": True,
-            "drinks": newDrinkSelection
-        })
+            # Get the updated drink selection
+            drinkSelection = Drink.query.order_by(Drink.id).all()
+            newDrinkSelection = []
 
-    except:
-        abort(400)
+            for drink in drinkSelection:
+                newDrinkSelection.append(drink.long())
+
+            return jsonify({
+                "success": True,
+                "drinks": newDrinkSelection
+            })
+
+        except:
+            abort(400)
 
 
 '''
@@ -159,40 +187,48 @@ def edit_drink(payload, drink_id):
         it should respond with a 404 error if <id> is not found
         it should delete the corresponding row for <id>
         it should require the 'delete:drinks' permission
-    returns status code 200 and json {"success": True, "delete": id} where id is the id of the deleted record
+    returns status code 200 and json {"success": True, "delete": id}
+    where id is the id of the deleted record
         or appropriate status code indicating reason for failure
 '''
+
+
 @app.route('/drinks/<int:drink_id>', methods=['DELETE'])
 @requires_auth('delete:drinks')
 def delete_drink(payload, drink_id):
-    try:
-        drink = Drink.query.get(drink_id)
+    if type(payload) is AuthError:
+        return auth_error(payload)
+    else:
+        try:
+            drink = Drink.query.get(drink_id)
 
-        if not drink:
-            abort(404)
+            if not drink:
+                abort(404)
 
-        # Get the json body to update the drink
-        drink.delete()
+            # Get the json body to update the drink
+            drink.delete()
 
-        # Get the updated drink selection
-        drinkSelection = Drink.query.order_by(Drink.id).all()
-        drinkSelection = []
+            # Get the updated drink selection
+            drinkSelection = Drink.query.order_by(Drink.id).all()
+            drinkSelection = []
 
-        for drink in drinkSelection:
-            drinkSelection.append(drink.long())
+            for drink in drinkSelection:
+                drinkSelection.append(drink.long())
 
-        return jsonify({
-            "success": True,
-            "drinks": drinkSelection
-        })
+            return jsonify({
+                "success": True,
+                "drinks": drinkSelection
+            })
 
-    except:
-        abort(400)
+        except:
+            abort(400)
 
 # Error Handling
 '''
 Example error handling for unprocessable entity
 '''
+
+
 @app.errorhandler(422)
 def unprocessable(error):
     return jsonify({
@@ -212,6 +248,8 @@ def unprocessable(error):
                     }), 404
 
 '''
+
+
 @app.errorhandler(400)
 def bad_request(error):
     return jsonify({
@@ -219,6 +257,7 @@ def bad_request(error):
         "error": 400,
         "message": "Bad Request."
     }, 400)
+
 
 @app.errorhandler(401)
 def unauthorized(error):
@@ -228,6 +267,7 @@ def unauthorized(error):
         "message": "Unauthorized."
     }, 401)
 
+
 @app.errorhandler(403)
 def forbidden_error(error):
     return jsonify({
@@ -236,14 +276,17 @@ def forbidden_error(error):
         "message": "Content Forbidden."
     }, 403)
 
+
 '''
 @TODO implement error handler for 404
     error handler should conform to general task above
 '''
+
+
 @app.errorhandler(404)
 def no_resource(error):
     return jsonify({
-        "success": False, 
+        "success": False,
         "error": 404,
         "message": "Resource not found"
     }), 404
@@ -252,10 +295,12 @@ def no_resource(error):
 @TODO implement error handler for AuthError
     error handler should conform to general task above
 '''
-# @app.errorhandler(403)
-# def auth_error(error):
-#     return jsonify({
-#         "success": False
-#         "code": "authorization_header_missing",
-#         "description": "Authorization header is expected."
-#     }, 403)
+
+
+@app.errorhandler(AuthError)
+def auth_error(authError):
+    return jsonify({
+        "success": False,
+        "error": authError.status_code,
+        "message": authError.error
+    }), authError.status_code
